Fix bug that occurs when checking/creating folders with multiple workers
When running with multiple workers enabled, it can happen that
the folder is created by  another worker between checking for
folder existence and creating new folder. If exists_ok is set to False
the application will return an error.
This fix changes the _get_file_path function so that it does not throw
and error if a folder that we are creating already exists.