Add fail_on_missing flag to ExportToTiff task.
Adds fail_on_missing flag to ExportToTiff task which determines whether
the task should fail if an export of a missing feature is requested.

If fail_on_missing is set to True -  a ValueError will be raised when
trying to Export a feature that is not present in the eopatch.

If fail_on_missing is set to False - only a warning will be logged

By default, the flag is set to True.